2020-04-02-website-colors/requirements.txt
pandas==1.0.3
Pillow==7.1.0
selenium==3.141.0
